:truck: Rename PaymentActions to Payments
