+++ frontend/src/pages/ProjectDetails/Payments/PaymentForm/WorkEstimation/BudgetBar/BudgetBar.stories.tsx
