+++ frontend/src/pages/ProjectDetails/Payments/PaymentForm/WorkEstimation/BudgetBar/index.tsx
