+++ frontend/src/pages/ProjectDetails/Payments/RemainingBudget/RemainingBudget.stories.tsx
